Add a compact binary encoding for the import section
The import section of a wasm binary must repeat the first name for each
import (the "module name"). This can be costly if there are many imports
in the same module, which is common. This patch adds a new encoding for
the import section in which the module name occurs only once.

This new encoding is disambiguated from previous encodings by means of
an invalid UTF-8 `name` consisting of a single 0xFF byte. 0xFF is never
valid in UTF-8; therefore, compliant implementations today should reject
the new encoding as having an invalid module name.